add support for natively-assigned "$oid" hex strings, in addition to user-specified IDs
I just figured out that natively-assigned $oid values to the "_id"
property actually encode the created-timestamp [1] of the created
document.

The ObjectId constructor throws an exception if the provided argument
is not a mongo OID value.  With this change, we can tell the
difference.  If the ObjectID constructor throws an exception, then the
ID filter will be _id: "<user-supplied-id>" as before, but if it
DOESN'T throw an exception, then we used the ObjectId wrapper, and the
condition ends up being ._id : { $oid: "native-oid-hex-string" }'

[1] http://docs.mongodb.org/manual/reference/object-id/